[COZY-177] refactor: FCM 패키지 위치 global -> domain 변경
