+++ src/main/java/com/cozymate/cozymate_server/domain/fcm/NotificationContentDto.java
@@ -1,4 +1,4 @@
-package com.cozymate.cozymate_server.global.fcm;
+package com.cozymate.cozymate_server.domain.fcm;
 
 import com.cozymate.cozymate_server.domain.member.Member;
 import java.util.List;
