+++ src/main/java/com/cozymate/cozymate_server/domain/fcm/NotificationService.java
@@ -1,16 +1,15 @@
-package com.cozymate.cozymate_server.global.fcm;
+package com.cozymate.cozymate_server.domain.fcm;
 
 import com.cozymate.cozymate_server.domain.member.Member;
 import com.cozymate.cozymate_server.domain.notificationlog.NotificationLog;
 import com.cozymate.cozymate_server.domain.notificationlog.enums.NotificationType;
 import com.cozymate.cozymate_server.domain.notificationlog.repository.NotificationLogRepository;
-import com.cozymate.cozymate_server.global.fcm.NotificationTargetDto.GroupTargetDto;
-import com.cozymate.cozymate_server.global.fcm.NotificationTargetDto.OneTargetReverseDto;
-import com.cozymate.cozymate_server.global.fcm.NotificationTargetDto.OneTargetDto;
+import com.cozymate.cozymate_server.domain.fcm.NotificationTargetDto.GroupTargetDto;
+import com.cozymate.cozymate_server.domain.fcm.NotificationTargetDto.OneTargetReverseDto;
+import com.cozymate.cozymate_server.domain.fcm.NotificationTargetDto.OneTargetDto;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.google.firebase.messaging.Message;
-import com.google.firebase.messaging.Notification;
 import java.util.HashMap;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
