+++ src/main/java/com/cozymate/cozymate_server/global/scheduler/NotificationScheduler.java
@@ -6,8 +6,8 @@ import com.cozymate.cozymate_server.domain.member.Member;
 import com.cozymate.cozymate_server.domain.notificationlog.enums.NotificationType;
 import com.cozymate.cozymate_server.domain.todo.Todo;
 import com.cozymate.cozymate_server.domain.todo.repository.TodoRepository;
-import com.cozymate.cozymate_server.global.fcm.NotificationService;
-import com.cozymate.cozymate_server.global.fcm.NotificationTargetDto.OneTargetDto;
+import com.cozymate.cozymate_server.domain.fcm.NotificationService;
+import com.cozymate.cozymate_server.domain.fcm.NotificationTargetDto.OneTargetDto;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
